Initial commit of files
Currently the sketch will correctly process button pressing from two buttons, but adding additional buttons into the code is trivial. I have added a few notes as to what I think I can do to improve this code. This is the first time for me working with the Arduino outside of early tutorials so I am still unfamiliar with the language.