Fixed initial commit
Initial commit was uploaded using a file downloaded directly from Arduino Create. Apparently the file is all in one line, which is gross. This commit also contains the source code to power the LEDS and the code to convert the button presses into Keyboard events. One small issue is that the Arduino UNO does not directly support using the Keyboard library, so I have commented out the code for now so that the project will compile. I have also placed an order for an Arduino Leonardo so this project will be updated to use that device.