Removed some of the dead code, also added support for the ASCII KeyCode for ArcadeButton
Removed the code that is not being used anymore for button presses. I left the SendToKeyboard function alone as I will need to rework that logic once I get the new board in. Also added support to send the ASCII Key Code to the ArcadeButton, this will come into usage once commands are sent back to the computer.