Fixed compilation issues, added support for LED library
I got the new Leonardo board and I noticed that the LEDs were not blinking correctly. Now this could be due to using the LED library but I would be surprised if using an object would effect performance that much. Originally on the Uno the LED required ten million cycles to flash the buld almost once a second, running on the Leonardo it requires three thousand to do the same. That marks a drastic change to the performance, which at this point is something interesting more than a blocker.