Add support for Timer library
I have added minimal support for the new Timer library, this was really just to debug the new source code. From here I will need to build a mechanism that will watch the remaining time that is passed back by UpdateTimer and to see if that number is larger compared to the last clock cycle. When that scenario occurs we can consider the timer to have gone off, we then must reset the alarm and then repeat the process.