Button and LED logic finalized.
Removed all timers from the sketch, I was not overly thrilled with the LEDs just blinking. The LEDs will now be on constantly except for when the button is depressed, then the LED will turn off.
Buttons and LEDs are now created as part of arrays of their respective objects.
Button inputs are now monitored using interrupts, this led to issues with needing to debounce the button. Tried several attempts at the software and nothing was 100% accurate so I chose to debounce the switches using a Schmitt Trigger in the hardware.