Refactored to use one interrupt
I have had lots of issues on the hardware side of this project to use all of the built in interrupts on the Leonardo. Using pins 0 and 1 have caused issues where a lot of false interrupts were being triggered. I think that could be caused by the fact that the 0 and 1 pins are being used for the serial communication. Realized that I could simplify the software by having all of the hardware buttons trigger the same interrupt pin, 3 in the code now, and then each button could maintain its position, pressed or released, with digital pins assigned directly to the buttons, pins 4 - 7 in the code.